Rename action_consts.js to action_types.js
