+++ botReducer.js
@@ -5,7 +5,7 @@
 
 'use strict';
 const { Markup } = require('telegraf');
-const { ACTION_HEARS_CANCEL, ACTION_HEARS_DONE, ACTION_HEARS_RESULTS } = require('./action_consts');
+const { ACTION_HEARS_CANCEL, ACTION_HEARS_DONE, ACTION_HEARS_RESULTS } = require('./action_types');
 const createDBStorage = require('./storage');
 const storage = createDBStorage();
