+++ createStore.js
@@ -1,5 +1,5 @@
 'use strict';
-const { ACTION_HEARS_CANCEL, ACTION_HEARS_DONE, ACTION_HEARS_RESULTS } = require('./action_consts');
+const { ACTION_HEARS_CANCEL, ACTION_HEARS_DONE, ACTION_HEARS_RESULTS } = require('./action_types');
 
 function createStore(storeReducer, storeState = {}) {
   let state = storeState;
